double_precision

+++ qed_fermion/hmc_sampler_batch.py
@@ -21,11 +21,12 @@ sys.path.insert(0, script_path + '/../')
 from qed_fermion.utils.coupling_mat3 import initialize_coupling_mat3, initialize_curl_mat
 
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+# device = torch.device('cpu')
 print(f"device: {device}")
 
-dtype = torch.float32
-cdtype = torch.complex64
-cg_dtype = torch.complex64
+dtype = torch.float64
+cdtype = torch.complex128
+cg_dtype = torch.complex128
 print(f"dtype: {dtype}")
 print(f"cdtype: {cdtype}")
 print(f"cg_cdtype: {cg_dtype}")
